Added load following (V2H/S2V)
Very scrappy code right now, but it does work as a proof of concept.

loadFollower.py polls once per second and sets the EVSE to match the grid power within 1A.

I've started abstracting the libraries for the Wallbox and the Shelly which is essential for looking at other hardware.

I'm probably going to look at moving to .NET at this point in a different repo. This ideally needs to be programmed using a cross-platform framework. Possibly MAUI for the user interface support. Java might work well for embedded platforms and is another possible way forward. There appear to exist Modbus libraries for both.