Added minimum and maximum charge levels
When performing load following, there may be times when it might only be desirable to charge or discharge, and there might be times when it is best to keep charging or discharging at 3A (with the wallbox quasar) so that it can respond rapidly to a change in solar energy. Minimum and maximum charging currents can now be specified in the load follower.

Also I've added a slight optimisation to the time from start-up to charging, but this might be specific to my wallbox (as might be the case for all of the guard times).

The load following functionality still needs to be moved to the library, but it would be possible to implement everything that is required just by using the load follower and specifying equal minimum and maximum charging levels.