Thread-based controller added
A thread-based controller that takes the desired mode of charging and implements it has been added. Scheduler2.py makes use of it; that can be used as an example. There are several modes of charging as follows:

* DORMANT: quiescent state (no charging or discharging happening)
* FULL_CHARGE: charging at the maximum possible power level (the "typical" car charging state)
* FULL_DISCHARGE: discharging at the maximum possible power level (V2G)
* LOAD_FOLLOW_CHARGE: allow charging to follow the grid (useful for using solar energy, sometimes known as S2V)
* LOAD_FOLLOW_DISCHARGE: allow discharging to follow the grid (useful for powering the home from the EV, sometimes known as V2H)
* LOAD_FOLLOW_BIDIRECTIONAL: allow charging and discharging to follow the grid (i.e. combined S2V and V2H operation)

If the controller recognises that the EVSE is in the wrong mode, it will try to correct that periodically.

Logging has been implemented in a directory that must be present called log; the directory is in the .gitignore file. Logging is rather verbose at this time.