Simplify the flux.py Octopus Flux example
I noticed while observing the graphs that there were some suboptimal situations with the previous set of rules, including charging in the "Day" period to discharge in the "Peak" period; given the efficiency of the charge-discharge cycle is likely around 80%, this was actually costing me money. So I've come up with what I believe is a sensible, much shorter set of rules. This will also be easier to understand for anyone who wants to implement their own driver for a different tariff.