Better treatment over SoC limits
I'm using 20% as my lower SoC limit (obviously that can be changed) and 97% as my upper SoC limit (which can also be changed but 97% is the limit that Quasar Wallbox will usually charge to). I've adjusted the code to document that. I've also changed the code to prevent logging of an intention to charge beyond the upper SoC limit and to stop resetting the state change timeout when that happens.