More cleanup and fixes
New getEvseCurrent() method in EvseInterface and implemented in WallboxQuasar to read the EVSE current. Returns the value of the register if charging is active and 0 if charging is paused.

Removed calcGridPower from EvseInterface and the WallboxQuasar implementation as the grid power comes from the Power interface and Shelly implementation and has nothing to do with the EVSE implementation. This was bad design and a hangover from the early days of this project.

EvseController now uses actual EVSE power reports from the current transformer of the Shelly as opposed to the WallboxQuasar getEvseCurrent() method but a comment is in place telling you what you need to replace if you want to use Shelly channel 1 for something other than the EVSE power.