Minor changes:
I didn't want the Flux discharge to take the power much below 30%. In my region at this time, it doesn't make sense to sell energy during the peak rate and then buy it back at the day rate. So it's better to buy the energy during the flux period and buy the solar energy and just sell that. If you're using more energy, it's better for it to come from the battery than for you to buy it from the grid. So it does the following now:

(a) If SoC is below 31%, do load follow discharge, allowed from 3A to the highest possible. Stop discharging if there's not enough home demand.
(b) If SoC is between 31% and 50% inclusive, discharge at 10A and allow it to rise if needed to cover home load.
(c) If SoC is greater than 50%, discharge at the maximum rate.

I made state change checks a little more frequent at every 20 seconds rather than every 30.

The timeout when talking Modbus to the Wallbox Quasar has been set to five seconds; the default is 30. This improves the response if there is a communications problem.

Also the CHANGE_SoC entry now includes the old SoC, the new SoC and the time between the old state and the new state.

I've added placeholders for Octopus account number and API keys. I want to create a system that makes a decision tree in the future that will automatically decide on the optimum state for the charger to be in (whether it should charge at the full rate, discharge at the full rate, mix charging or discharging with load following, and all that sort of stuff) which I'm planning to do by using the concept of a decision tree and minimax algorithm.

Also, you are now advised to make a secret.py file containing your configuration. If that's present, it will be picked up by configuration.py. If you use secret.py, you will not be invited to check it into your own repo.