Minor update to ignore reported SoC of 4%
I don't know why this happens but the Wallbox Quasar sometimes reports a SoC of 4% when that is not valid. It may well be that this actually is a valid value under some circumstances for the SoC but I've never seen that. (I don't commonly let my battery charge level go that low and it would be in the least awkward to test, not to mention costly.)