Removed thread locking in WallboxQuasar.py and documented that the class is not thread-safe. The implementation of locking was causing mutex locks that would not break.
Noted that lastTargetCurrent was being set in the wrong location within the calculateTargetCurrent function in EvseController and fixed. This is part of the code that deals with hysteresis.

Changed logging of state changes such that they only happen if the state has in fact changed.

Added more details to README.md about setup. Removed section on deprecated code which has now been deleted. Added development environment notes so that people can avoid Pylance error reports.